pyproject.toml
[project]
name = "llm-from-scratch"
version = "0.1.0"
description = "Building Large Language Models from scratch"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel",
    "tiktoken",
    "ruff",
    "pyyaml",
    "structlog",
    "numpy",
    "torch",
    "matplotlib",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "types-google-cloud-ndb",
    "types-requests",
    "google-auth-stubs",
    "tqdm",
    "tensorboard",
    "datasets",
    "tokenizers",
    "ollama",
]
